fixing the problem of having level-core and level_core; (this pr is just for the migration)
