+++ level_core/datastore/firestore/exceptions.py
